+++ lib/logstash-logger/device.rb
@@ -5,9 +5,11 @@ module LogStashLogger
     DEFAULT_TYPE = :udp
 
     autoload :Base, 'logstash-logger/device/base'
+    autoload :Connectable, 'logstash-logger/device/connectable'
     autoload :Socket, 'logstash-logger/device/socket'
     autoload :UDP, 'logstash-logger/device/udp'
     autoload :TCP, 'logstash-logger/device/tcp'
+    autoload :Redis, 'logstash-logger/device/redis'
     autoload :File, 'logstash-logger/device/file'
     autoload :Stdout, 'logstash-logger/device/stdout'
 
@@ -22,6 +24,7 @@ module LogStashLogger
         when :udp then UDP
         when :tcp then TCP
         when :file then File
+        when :redis then Redis
         when :stdout then Stdout
         else fail ArgumentError, 'Invalid type'
       end