+++ spec/spec_helper.rb
@@ -41,4 +41,6 @@ RSpec.shared_context 'device' do
 
   let(:file) { Tempfile.new('test') }
   let(:file_device) { LogStashLogger::Device.new(type: :file, path: file.path)}
+
+  let(:redis_device) { LogStashLogger::Device.new(type: :redis) }
 end